Make fscache be a mount option for Ceph (like NFS)
Changed it so fscache is not on by default. Instead it requires a mount flag of
"fsc". This will allow people who build distro kernels to enable caching but
still leave the decision to use it up to the user. This mirrors how NFS behaves
with the fscache mount options.

Signed-off-by: Milosz Tanski <mtanski@adfin.com>