vfs: RWF_NOWAIT flag for preadv2
generic_file_read_iter() supports a new flag RWF_NOWAIT to indicates that
want to read the data only if it's already in the page cache.

Additionally, there are a few filesystems that we have to specifically
bail early if RWF_NOWAIT because the op would block. Christoph Hellwig
contributed this code.

Signed-off-by: Milosz Tanski <milosz@adfin.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jeff Moyer <jmoyer@redhat.com>
Acked-by: Sage Weil <sage@redhat.com>